requirements.txt
Django
djangorestframework
django-rest-knox
coverage
gunicorn
django-heroku